front-back productdetails

+++ RealEstateBackend/Models/DTOs/ReviewDto/ReviewDto.cs
@@ -13,7 +13,7 @@ namespace RealEstate.Models.DTOs.ReviewDto
         public int ProductId { get; set; }
 
         [Range(1, 5)]
-        public int Rating { get; set; }
+        public double Rating { get; set; }
 
         [MaxLength(300)]
         public string? Comment { get; set; }
